pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "forex-signal-bot"
version = "0.1.0"
description = "Forex signal alert bot for email notifications."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pandas>=2.2.2",
  "numpy>=1.26.4",
  "requests>=2.32.3",
  "ta>=0.11.0",
  "APScheduler>=3.10.4",
  "python-dotenv>=1.0.1"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
